Weekly Update
- Take out PAUSE BUTTON & LOG
-  Replace PAUSE with Q-OUT / STOP in Queue list
- Check Q-OUT / STOP status in directive-while-loop
- Take out REPEAT BUTTON & LOG
- Replace it with REPETITIVE WAVEFORM instead and limit progress within 100%
- Take out run_status & Move it  to Queue list page
- Check instrument availability before starting measurement to avoid conflict in connection
- Add REQUEUE (DATA=0) button under DATA column
- Remember BOTH user's last queue and sample
- Run clearance based on the timely sample open for measurement instead

+++ TEST/FACE/pyqum/instrument/logger.py
@@ -142,14 +142,15 @@ class address:
     Set <reset=False> to directly load from LOG if it contains "address" 
     '''
     def __init__(self):
-        with open(ADDRESS_PATH / "address.json") as ad:
-            self.book = json.load(ad)
+        # with open(ADDRESS_PATH / "address.json") as ad:
+        #     self.book = json.load(ad)
+        self.db = get_db()
         
     def lookup(self, instr_name, label=1):
         '''Lookup from the database instead from the book'''
         
         self.instr_name = instr_name
-        self.rs = get_db().execute(
+        self.rs = self.db.execute(
             'SELECT m.address\n' +
             'FROM machine m\n' +
             'WHERE m.codename = ?',
@@ -164,18 +165,16 @@ class address:
         Update SQL Database:
         connected: 0 or 1, codename = <instr>-<label/index> 
         '''
-        db = get_db()
-        db.execute( 'UPDATE machine SET user_id = ?, connected = ? WHERE codename = ?', (session['user_id'], connected, codename,) )
-        db.commit()
+        
+        self.db.execute( 'UPDATE machine SET user_id = ?, connected = ? WHERE codename = ?', (session['user_id'], connected, codename,) )
+        self.db.commit()
         return
-    # def machine_list(self,category):
-    #     machlist = get_db().execute(
-    #         'SELECT m.codename, m.connected\n' +  
-    #         'FROM machine m\n' +
-    #         'WHERE m.category = ?'
-    #         (category,)
-    #     ).fetchall()
-    #     return machlist
+    def macantouch(self,instr_list):
+        '''return total connection(s) based on instrument-list given'''
+        connection = 0
+        for mach in instr_list:
+            connection += int(self.db.execute('''SELECT connected FROM machine WHERE codename = ?''', (mach,) ).fetchone()['connected'])
+        return connection
 
 class specification:
     '''lookup specifications for each instruments
@@ -276,20 +275,19 @@ class measurement:
         2. Checking Database if any (daylist)
         3. Used for sending status to the front-end via JS
     '''
-    def __init__(self, mission, task, owner='USR', sample='Sample', loopcount=[], loop_dur=[]):
+    def __init__(self, mission, task, owner='USR', sample='Sample'):
         # Primary parameters (mission & task is auto-detected by OS)
         self.mission, self.task = mission, task
         self.owner, self.sample = owner, sample
         self.mssnpath = Path(USR_PATH) / owner / sample / mission
-        self.loopcount, self.loop_dur = loopcount, loop_dur # for testeach
         #current location
         self.place = ", ".join(location()) 
-        
+        self.status = "M INTIATED"
         # FOR Resume / Access operation:
         try:
             daylist = [d for d in listdir(self.mssnpath) if isdir(self.mssnpath / d)]
-            print("There are %s days" %len(daylist))
-            # filter out non task-specific
+            # print("There are %s days" %len(daylist))
+            # filter out non-task-related
             relatedays = []
             for d in daylist:
                 task_relevant_time = [t for t in listdir(self.mssnpath / d) if t.split('.')[0] == self.task]
@@ -322,7 +320,7 @@ class measurement:
         '''corder: {parameters: <waveform>}\n'''
 
         # New operation if "new" is selected:
-        if index < 0:
+        if index == -1:
             now = datetime.now() #current day & time
             self.day = now.strftime("%Y-%m-%d(%a)")
             self.moment = now.strftime("%H:%M:%f")
@@ -353,13 +351,19 @@ class measurement:
 
                     # Insert into Queue-list on SQL-Database: (from # 3. in settings)
                     jobstart(self.day, task_index, JOBID)
-
+                    self.status = "JOBID #%s STOPPED" %JOBID # By the time this is output, M has exitted
                     break
                 else:
                     task_index += 1
+
+        # LOG-TEMP if "temp" is selected:
+        elif index == -3:
+            '''PENDING'''
+            pass
+            
         
         # from database:
-        else:
+        elif index >= 0:
             try:
                 self.day = self.daylist[index]
                 self.taskentries = [int(t.split('(')[1][:-1]) for t in listdir(self.mssnpath / self.day) if t.split('.')[0] == self.task]
@@ -367,6 +371,8 @@ class measurement:
             except(ValueError): 
                 print("index might be out of range")
                 pass
+        
+        else: print(Fore.RED + "INVALID INDEX (%s) FOR DAY SELECT..." %index)
 
     # only for scripting
     def whichmoment(self):
@@ -570,34 +576,47 @@ def settings(datadensity=1):
     @wrapt.decorator
     def wrapper(Name, instance, a, b):
         Generator = Name(*a, **b)
-        owner, sample, tag, instr, corder, comment, dayindex, taskentry, testeach, perimeter, queue = next(Generator)
+        owner, sample, tag, instr, corder, comment, dayindex, taskentry, perimeter, queue = next(Generator)
         mission = Path(inspect.getfile(Name)).parts[-1].replace('.py','') #Path(inspect.stack()[1][1]).name.replace('.py','')
         task = Name.__name__
         # print("task: %s" %task)
         M = measurement(mission, task, owner, sample) #M-Initialization
         if type(dayindex) is str: # for later access
             pass # ONLY M-INITIALIZATION (everytime when click a task) for the LATTER data access
-        elif type(dayindex) is int: # for new, resume, test-loop/each
-            if testeach:
-                M.loopcount, M.loop_dur = next(Generator)
-            else:
+        elif type(dayindex) is int: # for temp (-3), new (-1), resume (>=0)
+            
+            if g.user['measurement']:
                 # 1. Register or Retrieve JOB(ID):
-                if dayindex < 0:
-                    JOBID = jobin(task, corder, perimeter, instr, comment, tag)
-                else: 
+                if dayindex == -1: # NEW FILE
+                    # REQUEUE from previous dropped out JOB => (0 file, 1 job)
+                    if 'jobid' in perimeter.keys(): JOBID = perimeter['jobid']
+                    # NEW JOB => (0 file, 0 job)
+                    else: JOBID = jobin(task, corder, perimeter, instr, comment, tag)
+                    print(Fore.BLUE + "NEW DAY DETECTED")
+                elif dayindex == -3: # TEMP FILE
+                    pass
+                elif dayindex >= 0: # RESUME from previous stopped File => (1 file, 1 job)
                     day = M.daylist[dayindex]
                     criteria = dict(samplename=sample, task=task, dateday=day, wmoment=taskentry)
                     JOBID = jobsearch(criteria)
+                    print(Fore.BLUE + "OLD DAY DETECTED")
+                else: print(Fore.RED + "INVALID DAYINDEX: %s" %dayindex)
+                perimeter['jobid'] = JOBID
 
                 # 2. Queue-IN and Wait for your turn:
-                qin(queue, JOBID)
+                M.status = qin(queue, JOBID)
                 while True:
                     lisqueue(queue)
+                    sleep(7)
                     if JOBID not in g.jobidlist[queue]: # get out in the middle of waiting
-                        return None
-                    elif g.jobidlist[queue].index(JOBID) == 0: # only first-in-line get to break the waiting loop
-                        set_status(task, dict(pause=False))
+                        return M
+                    elif g.jobidlist[queue].index(JOBID)==0 and not address().macantouch(list(instr.values())):
+                        '''All of the following should be fulfilled before taking turn to run:
+                            1. ONLY FIRST-IN-LINE get to break the waiting loop
+                            2. ALL instruments required are disconnected
+                        '''
                         break
+                    print(Fore.YELLOW + "JOBID #%s is waiting every 7 seconds" %JOBID)
 
                 # 3. Start RUNNING / WORKING / MEASUREMENT:
                 M.selectday(dayindex, corder, perimeter, instr, datadensity, comment, tag, JOBID)
@@ -611,6 +630,8 @@ def settings(datadensity=1):
                         # sleep(3) #for debugging purposes
                 except(KeyboardInterrupt): print(Fore.RED + "\nSTOPPED")
 
+            else: M.status = "M JOB REJECTED: PLS CHECK M-CLEARANCE!"
+
         # Measurement Object/Session:
         return M
     return wrapper
@@ -623,11 +644,11 @@ def lisample(usr):
 def lisjob(usr, sample, queue, maxlist=12):
     '''list jobs for queue-page under MSSN'''
     # Provide user's clearances for each Queue (CHAR0, QPC0):
-    if g.user['instrument'] and g.user['measurement']:
+    if g.user['measurement']:
         # Extracting list from SQL-Database:
         Joblist = get_db().execute(
             '''
-            SELECT j.id, j.task, j.dateday, j.wmoment, j.startime, j.instrument, j.comment
+            SELECT j.id, j.task, j.dateday, j.wmoment, j.startime, j.instrument, j.comment, j.progress
             FROM user u
             INNER JOIN job j ON j.user_id = u.id
             INNER JOIN sample s ON s.id = j.sample_id
@@ -644,8 +665,7 @@ def lisqueue(queue):
     list queues for queue-page under MSSN
     Update clearance for running the experiment
     '''
-    session['run_clearance'] = False
-    if g.user['instrument'] and g.user['measurement']:
+    if g.user['measurement']:
         try:
             g.Queue, g.jobidlist = {}, {}
 
@@ -663,15 +683,6 @@ def lisqueue(queue):
             g.Queue[queue] = [dict(x) for x in g.Queue[queue]]
             g.jobidlist[queue] = [x['id'] for x in g.Queue[queue]] # use to scheduling tasks in queue
         
-            # To queue-in, sample chosen (MEAL) must be aligned with the assigned sample for that queue-system (by admin):
-            asample = get_db().execute(
-                '''
-                SELECT samplename
-                FROM queue
-                WHERE system = ?
-                ''', (queue,)
-                ).fetchone()['samplename'] # assigned sample by admin
-            session['run_clearance'] = bool( asample==get_status("MSSN")[session['user_name']]['sample'] )
         except: pass
         # print(Fore.BLACK + Back.WHITE + "Clearance for queue %s: %s"%(queue, session['run_clearance']))
     return
@@ -679,23 +690,28 @@ def lisqueue(queue):
 # QUEUE
 def qin(queue,jobid):
     '''Queue in with a Job'''
-    try:
-        db = get_db()
-        db.execute('INSERT INTO %s (job_id) VALUES (%s)' %(queue,jobid))
-        db.commit()
-        status = "Queued-in successfully"
-    except:
-        status = "Error"
+    if g.user['measurement']:
+        try:
+            db = get_db()
+            db.execute('INSERT INTO %s (job_id) VALUES (%s)' %(queue,jobid))
+            db.commit()
+            status = "Queued-in successfully"
+        except:
+            status = "Error Queueing in with JOBID #%s" %jobid
+    else: status = "Measurement clearance was not found"
     return status
 def qout(queue,jobid):
     '''Queue out without a Job'''
-    try:
-        db = get_db()
-        db.execute('DELETE FROM %s WHERE job_id = %s' %(queue,jobid))
-        db.commit()
-        status = "Queued-out successfully"
-    except:
-        status = "Error"
+    if g.user['measurement']:
+        try:
+            db = get_db()
+            db.execute('DELETE FROM %s WHERE job_id = %s' %(queue,jobid))
+            db.commit()
+            status = "JOBID #%s Queued-out successfully" %jobid
+        except:
+            raise
+            status = "Error Queueing out with JOBID #%s" %jobid
+    else: status = "Measurement clearance was not found"
     return status
 def qid(queue,jobid):
     '''Get queue number'''
@@ -708,40 +724,52 @@ def qid(queue,jobid):
 # JOB
 def jobin(task,corder,perimeter,instr,comment,tag):
     '''Register a JOB and get the ID for queue-in later while leaving day and task# blank first'''
-    try:
-        db = get_db()
-        samplename = get_status("MSSN")[session['user_name']]['sample']
-        queue = get_status("MSSN")[session['user_name']]['queue']
-        sample_id = db.execute('SELECT s.id FROM sample s WHERE s.samplename = ?', (samplename,)).fetchone()[0]
-        cursor = db.execute('INSERT INTO job (user_id, sample_id, task, parameter, perimeter, instrument, comment, tag, queue) VALUES (?,?,?,?,?,?,?,?,?)', 
-                                    (g.user['id'],sample_id,task,str(corder),str(perimeter),str(instr),comment,tag,queue))
-        JOBID = cursor.lastrowid
-        db.commit()
-        print(Fore.GREEN + Back.WHITE + "Successfully register the data into SQL Database with JOBID: %s" %JOBID)
-    except:
-        JOBID = None 
-        print(Fore.RED + Back.WHITE + "Check all database input parameters")
-        raise
+    if g.user['measurement']:
+        try:
+            db = get_db()
+            samplename = get_status("MSSN")[session['user_name']]['sample']
+            queue = get_status("MSSN")[session['user_name']]['queue']
+            sample_id = db.execute('SELECT s.id FROM sample s WHERE s.samplename = ?', (samplename,)).fetchone()[0]
+            cursor = db.execute('INSERT INTO job (user_id, sample_id, task, parameter, perimeter, instrument, comment, tag, queue) VALUES (?,?,?,?,?,?,?,?,?)', 
+                                        (g.user['id'],sample_id,task,str(corder),str(perimeter),str(instr),comment,tag,queue))
+            JOBID = cursor.lastrowid
+            db.commit()
+            perimeter['jobid'] = JOBID
+            sleep(1)
+            db.execute('UPDATE job SET perimeter = ? WHERE id = ?', (str(perimeter),JOBID))
+            db.commit()
+            print(Fore.GREEN + Back.WHITE + "Successfully register the data into SQL Database with JOBID: %s" %JOBID)
+        except:
+            # raise
+            JOBID = None 
+            print(Fore.RED + Back.WHITE + "Check all database input parameters")
+    else: JOBID = None
     return JOBID
 def jobstart(day,task_index,JOBID):
     '''Start a JOB by logging day and task#'''
-    try:
-        db = get_db()
-        db.execute('UPDATE job SET dateday = ?, wmoment = ? WHERE id = ?', (day,task_index,JOBID))
-        db.commit()
-        print(Fore.GREEN + Back.WHITE + "Successfully update JOB#%s with (Day: %s, TASK#: %s" %(JOBID,day,task_index))
-    except:
-        print(Fore.RED + Back.WHITE + "INVALID JOBID")
-        raise
+    if g.user['measurement']:
+        try:
+            db = get_db()
+            db.execute('UPDATE job SET dateday = ?, wmoment = ? WHERE id = ?', (day,task_index,JOBID))
+            db.commit()
+            print(Fore.GREEN + Back.WHITE + "Successfully update JOB#%s with (Day: %s, TASK#: %s" %(JOBID,day,task_index))
+        except:
+            print(Fore.RED + Back.WHITE + "INVALID JOBID")
+            raise
+    else: pass
     return
 def jobnote():
     '''Add NOTE to a JOB after analyzing the data'''
 
     return
 def jobsearch(criteria, mode='jobid'):
-    '''Search for JOB(s) based on criteria (keywords)'''
+    '''Search for JOB(s) based on criteria (keywords)
+        mode <jobid>: get job-id based on criteria
+        mode <tdmq>: get task, dateday, wmoment & queue based on job-id given as criteria
+    '''
     db = get_db()
     if mode=='jobid':
+        # as single-value
         result = db.execute(
                     '''
                     SELECT j.id 
@@ -750,6 +778,9 @@ def jobsearch(criteria, mode='jobid'):
                     WHERE s.samplename = ? AND j.task = ? AND j.dateday = ? AND j.wmoment = ?
                     ''', (criteria['samplename'], criteria['task'], criteria['dateday'], criteria['wmoment'])
                 ).fetchone()[0]
+    elif mode=='tdm':
+        # as dictionary
+        result = db.execute('SELECT task, dateday, wmoment, queue FROM job WHERE id = ?', (criteria,)).fetchone()
     else: # PENDING: for other kind of search
         result = None 
     return result
